Notifying Users By E-mail Their Password Is Going To Expire
Around 2009 I wrote a blog post about and also wrote a tool to notify users through e-mail when their password was going to expire. You can read all the details about the idea here. Now that tool was very inflexible and because of that I received numerous requests to make it more flexible such as the ability to customize the e-mail message. With this blog post I'm sharing a brand new tool, based upon PowerShell, that will notify users through e-mail when their password is going to expire. So let's get started in explaining on this works! I did not test all combinations! However, I do expect it to run on any Windows version as long as PowerShell is available. It should also work against any AD version and there is NO dependency on using the AD PowerShell CMDlets!. Everything is done through ADSI to be independent of Windows versions! It will also support PSOs if the DFL is high enough and PSOs are configured!

For a detailed explanation of this script, please go to:

(2016-05-09) Notifying Users By E-mail Their Password Is Going To Expire (Update 5)