Testing SYSVOL Replication Latency/Convergence Through PowerShell
A Microsoft PFE, that probable is using the this script, contacted me asking if I also had a version that could be used to test the latency/convergence of the SYSVOL replication. It was not that difficult to change the AD version in a way to check the SYSVOL replication. You might think I now have one script that does both AD and SYSVOL, but au contraire, there are two scripts. The one above does AD, and the one below does SYSVOL. The behavior/logic for the SYSVOL version of the script is very similar to the AD version of the script.

For a detailed explanation of this script, please go to:

(2013-03-03) Testing SYSVOL Replication Latency/Convergence Through PowerShell
(2014-02-17) Testing SYSVOL Replication Latency/Convergence Through PowerShell (Update 3)